requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.2
python-dotenv==1.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
whitenoise==6.7.0
